Move Radar functions out of component file
I could see these being split into separate files, but I think this is
enough for now.
