+++ plugins/tech-radar/src/components/Radar/Radar.tsx
@@ -15,152 +15,9 @@
  */
 
 import React, { FC, useState, useRef } from 'react';
-import { forceCollide, forceSimulation } from 'd3-force';
-import color from 'color';
 import RadarPlot from '../RadarPlot';
-import Segment from '../../utils/segment';
 import { Ring, Quadrant, Entry } from '../../utils/types';
-
-const adjustQuadrants = (
-  quadrants: Quadrant[],
-  radius: number,
-  width: number,
-  height: number,
-) => {
-  /*
-    0           1                             2           3 ← x stops index
-    │           │                             │           │ ↓ y stops index
-    ┼───────────┼─────────────────────────────┼───────────┼─0
-    │           │        . -- ~~~ -- .        │           │
-    │           │    .-~               ~-.    │           │
-    │    <Q3>   │   /                     \   │   <Q2>    │
-    │           │  /                       \  │           │
-    ┼───────────┼─────────────────────────────┼───────────┼─1
-    ┼───────────┼─────────────────────────────┼───────────┼─2
-    │           │ |                         | │           │
-    │           │  \                       /  │           │
-    │    <Q1>   │   \                     /   │   <Q0>    │
-    │           │    `-.               .-'    │           │
-    │           │        ~- . ___ . -~        │           │
-    ┼───────────┼─────────────────────────────┼───────────┼─3
-     */
-
-  const margin = 16;
-  const xStops = [
-    margin,
-    width / 2 - radius - margin,
-    width / 2 + radius + margin,
-    width - margin,
-  ];
-  const yStops = [margin, height / 2 - margin, height / 2, height - margin];
-
-  // The quadrant parameters correspond to Q[0..3] above.  They are in this order because of the
-  // original Zalando code; maybe we should refactor them to be in reverse order?
-  const legendParams = [
-    {
-      x: xStops[2],
-      y: yStops[2],
-      width: xStops[3] - xStops[2],
-      height: yStops[3] - yStops[2],
-    },
-    {
-      x: xStops[0],
-      y: yStops[2],
-      width: xStops[1] - xStops[0],
-      height: yStops[3] - yStops[2],
-    },
-    {
-      x: xStops[0],
-      y: yStops[0],
-      width: xStops[1] - xStops[0],
-      height: yStops[1] - yStops[0],
-    },
-    {
-      x: xStops[2],
-      y: yStops[0],
-      width: xStops[3] - xStops[2],
-      height: yStops[1] - yStops[0],
-    },
-  ];
-
-  quadrants.forEach((quadrant, idx) => {
-    const legendParam = legendParams[idx % 4];
-
-    quadrant.idx = idx;
-    quadrant.radialMin = (idx * Math.PI) / 2;
-    quadrant.radialMax = ((idx + 1) * Math.PI) / 2;
-    quadrant.offsetX = idx % 4 === 0 || idx % 4 === 3 ? 1 : -1;
-    quadrant.offsetY = idx % 4 === 0 || idx % 4 === 1 ? 1 : -1;
-    quadrant.legendX = legendParam.x;
-    quadrant.legendY = legendParam.y;
-    quadrant.legendWidth = legendParam.width;
-    quadrant.legendHeight = legendParam.height;
-  });
-};
-
-const adjustEntries = (
-  entries: Entry[],
-  activeEntry: Entry | null | undefined,
-  quadrants: Quadrant[],
-  rings: Ring[],
-  radius: number,
-) => {
-  let seed = 42;
-  entries.forEach((entry, idx) => {
-    const quadrant = quadrants.find(q => {
-      const match =
-        typeof entry.quadrant === 'object' ? entry.quadrant.id : entry.quadrant;
-      return q.id === match;
-    });
-    const ring = rings.find(r => {
-      const match = typeof entry.ring === 'object' ? entry.ring.id : entry.ring;
-      return r.id === match;
-    });
-
-    if (!quadrant) {
-      throw new Error(
-        `Unknown quadrant ${entry.quadrant} for entry ${entry.id}!`,
-      );
-    }
-    if (!ring) {
-      throw new Error(`Unknown ring ${entry.ring} for entry ${entry.id}!`);
-    }
-
-    entry.idx = idx;
-    entry.quadrant = quadrant;
-    entry.ring = ring;
-    entry.segment = new Segment(quadrant, ring, radius, () => seed++);
-    const point = entry.segment.random();
-    entry.x = point.x;
-    entry.y = point.y;
-    entry.active = activeEntry ? entry.id === activeEntry.id : false;
-    entry.color = entry.active
-      ? entry.ring.color
-      : color(entry.ring.color).desaturate(0.5).lighten(0.1).string();
-  });
-
-  const simulation = forceSimulation()
-    .nodes(entries)
-    .velocityDecay(0.19)
-    .force('collision', forceCollide().radius(12).strength(0.85))
-    .stop();
-
-  for (
-    let i = 0,
-      n = Math.ceil(
-        Math.log(simulation.alphaMin()) / Math.log(1 - simulation.alphaDecay()),
-      );
-    i < n;
-    ++i
-  ) {
-    simulation.tick();
-
-    for (const entry of entries) {
-      entry.x = entry.segment.clipx(entry);
-      entry.y = entry.segment.clipy(entry);
-    }
-  }
-};
+import { adjustQuadrants, adjustRings, adjustEntries } from './utils';
 
 type Props = {
   width: number;
@@ -172,23 +29,13 @@ type Props = {
 };
 
 const Radar: FC<Props> = props => {
-  const [activeEntry, setActiveEntry] = useState<Entry | null>();
-
-  const node = useRef<SVGSVGElement>(null);
-
-  const adjustRings = (rings: Ring[], radius: number) => {
-    rings.forEach((ring, idx) => {
-      ring.idx = idx;
-      ring.outerRadius = ((idx + 2) / (rings.length + 1)) * radius;
-      ring.innerRadius =
-        ((idx === 0 ? 0 : idx + 1) / (rings.length + 1)) * radius;
-    });
-  };
-
-  // TODO(dflemstr): most of this method can be heavily memoized if performance becomes a problem
   const { width, height, quadrants, rings, entries } = props;
   const radius = Math.min(width, height) / 2;
 
+  const [activeEntry, setActiveEntry] = useState<Entry | null>();
+  const node = useRef<SVGSVGElement>(null);
+
+  // TODO(dflemstr): most of this can be heavily memoized if performance becomes a problem
   adjustQuadrants(quadrants, radius, width, height);
   adjustRings(rings, radius);
   adjustEntries(entries, activeEntry, quadrants, rings, radius);
